allow to input password interactively
