+++ src/Command/CreateUserCommand.php
@@ -13,14 +13,11 @@ use App\Entity\User;
 use App\User\UserService;
 use App\Utils\CommandStyle;
 use App\Validator\ValidationFailedException;
-use Symfony\Component\Console\Command\Command;
-use Symfony\Component\Console\Helper\QuestionHelper;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
-use Symfony\Component\Console\Question\Question;
 
-final class CreateUserCommand extends Command
+final class CreateUserCommand extends AbstractUserCommand
 {
     private $userService;
 
@@ -90,32 +87,5 @@ final class CreateUserCommand extends Command
         }
 
         return 0;
-    }
-
-    /**
-     * @param InputInterface $input
-     * @param OutputInterface $output
-     *
-     * @return string
-     */
-    protected function askForPassword(InputInterface $input, OutputInterface $output): string
-    {
-        /** @var QuestionHelper $helper */
-        $helper = $this->getHelper('question');
-
-        $passwordQuestion = new Question('Please enter the password: ');
-        $passwordQuestion->setHidden(true);
-        $passwordQuestion->setHiddenFallback(false);
-        $passwordQuestion->setValidator(function (?string $value) {
-            $password = trim($value);
-            if (empty($password)) {
-                throw new \Exception('The password may not be empty');
-            }
-
-            return $value;
-        });
-        $passwordQuestion->setMaxAttempts(3);
-
-        return $helper->ask($input, $output, $passwordQuestion);
     }
 }
